Rename modules for layer-specific tests, to clearly indicate which leyaer they test.
