+++ run_tests.py
@@ -9,18 +9,18 @@ from tudor import generate_app
 import tests
 from tests.conversions_tests import *
 from tests.convert_task_to_tag_tests import *
-from tests.db_loader_tests import *
+from tests.ll_db_loader_tests import *
 from tests.user_tests import *
-from tests.is_user_authorized_or_admin_tests import *
-from tests.sort_by_hierarchy_tests import *
-from tests.get_index_data_tests import *
-from tests.get_deadlines_data_tests import *
-from tests.create_new_task_tests import *
-from tests.task_set_done_tests import *
-from tests.task_set_deleted_tests import *
-from tests.task_unset_done_tests import *
-from tests.task_unset_deleted_tests import *
-from tests.do_reset_order_nums_tests import *
+from tests.ll_is_user_authorized_or_admin_tests import *
+from tests.ll_sort_by_hierarchy_tests import *
+from tests.ll_get_index_data_tests import *
+from tests.ll_get_deadlines_data_tests import *
+from tests.ll_create_new_task_tests import *
+from tests.ll_task_set_done_tests import *
+from tests.ll_task_set_deleted_tests import *
+from tests.ll_task_unset_done_tests import *
+from tests.ll_task_unset_deleted_tests import *
+from tests.ll_do_reset_order_nums_tests import *
 from tests.tag_tests import *
 from tests.task_tests import *
 from tests.task_id_tests import *
@@ -28,12 +28,12 @@ from tests.task_css_tests import *
 from tests.task_tags_tests import *
 from tests.task_duration_text_tests import *
 from tests.task_cost_text_tests import *
-from tests.search_tests import *
+from tests.ll_search_tests import *
 from tests.task_dependencies_tests import *
 from tests.ll_task_tags_tests import *
-from tests.get_lowest_highest_order_num import *
+from tests.ll_get_lowest_highest_order_num import *
 from tests.task_prioritize_tests import *
-from tests.route_tests import *
+from tests.vl_route_tests import *
 from tests.persistence_layer_tests import *
 from tests.ll_do_import_data_tests import *
 from tests.ll_set_task_tests import *
